decrease learning rate for interesting results
a learning rate of 0.02 was pretty good, but maybe too good. most of
the time the loss graph looked the same, except very rarely where it
got stuck on a local minimum.

a learning rate of 0.001 doesn't approach 0 that quickly, so the curve
looks a bit more interesting, and is more varied with each run.